+++ apps/hello_world/host/host_kernels.hpp
@@ -6,9 +6,9 @@ namespace cpu {
 
 namespace kernels {
 
-void vector_add(const int *input_a,
-                const int *input_b,
-                int *output,
+void vector_add(const float *input_a,
+                const float *input_b,
+                float *output,
                 size_t start,
                 size_t end);
